While closing beekeper, wait for the pid file to be removed
An issue was observer while running
`clive beekeeper spawn`
command right after closing beekeeper via the
testnet_node.py script.

In the logs it could be observed that:
```
Testnet node is ready to use
Beekeeper is already running on http://0.0.0.0:44009 with pid 318
```

So this commit should prevent such a situation happening again.