Fix the issue with infinite loading on Savings screen
This was caused by changed logic in DataProvider.
SavingsDataProvider instance was created with
paused=True when it shouldn't be.
Also it wa previously doing a init_update.

Now the default paused=False is to avoid such a situation in the future.
But it can be explicitly set by the initializer params.