Pause DataProvider when screen, that it's mounted on, gets suspended
Need of this behaviour comes from the fact
on_unmount is not called on DataProvider (or any other widget)
when some other screen is pushed on the screen, where
DataProvider (or any other widget) was already mounted.

And we don't want the dataprovider to get the data for screen
that's currently not on the top.