Fix pop_screen_unitl posting ScreenResume message too early
This caused some issues with providers when
fast broadast was used on the operatin summary screen
and self.provider.content was referenced.
That's because of screen resume was posted during each pop_screen
and looks like provider catched that ScreenResume after screen yielding
that provider was acutally popped (so NoMatches occured when querying
provider which notified about change).