Call first_try_callback with parameters
In some cases, it happened that the parameters old_vaue and value
were passed as None, and in the meantime the data arrived. So when the
calling `first_try_callback` the result was positive, which was true
because the data had arrived. To avoid this situation, we pass
interested data as an argument to the function that checks whether the data is available
(first_try_callback), so the values and the check callback work in the same
state of the data.