Avoid the situation where an account is not found by some acount referencing widget
This could be a situation where we remove some account from, for example, watched account and the widget referring to that account asks for some data, but the account no longer exists. This happens because the dashboard, which also observes profile_data, has not yet removed this widget. This try-except block avoids such situations