Remove Commands.get_config as its unused and there is no need for its existance
Config is stored in Node.CachedData and is
updated with every node address change.

Since node shouldn't (and I think also can't) update it's
config when it's running, we don't need ever to ask for a "new" config
via Commands, and can refer to the cached value only.

Note:
Commands.get_dynamic_global_properties should still exist in
this place, because in case of DGPO  is common that we may want
to get the latest data and at the same time update the cache
(via call to Commands.get_dynamic_global_properties)
or just get the currently stored/cached data
(via call to Node.cached.dynamic_global_properties).
That's a local decision to use cached gdpo or get a recently one.