Fix run_instance.sh terminal resizing issue
When common.sh is included, when starting clive,
`tput lines` and `tput cols` result with a static
value of terminal size.

There was a workaround so instead of 80x24 during startup,
it was resized to valid size due to `--env` LINES and `--env COLUMNS`,
but when resizing terminal window, the size is not updated this way.

Since common.sh is not used in this script, better not to
source it. So we can have correctly working resizing.