Longer timeout value instead of multiple retries with shorter timeout
Because it looks there is no sense to retry on timeout, just have a
longer timeout. That's because if something is under heavy load, and we
have a timeout set for e.g. 6s while it potentially could answer in 7s,
we won't be able to correctly communicate and still
CommunicationTimeoutError will be send after the same time as longer
timeout set (6s * 5 (retries) = 30s)

So with this change we should have less timeout errors and the time for
TimeoutError notification to display will remain the same.