Improve the way header data is updated (block, online status, network type)
Previously due to the usage of async methods,
which have a asyncio.Lock inside,
there was a issue because one widget hold the lock
when there was no connection with a node.
Other widgets displayed "loading..." for a long time.

Now all node-related header data is updated in the same moment, asap.

Also improved the situation where Node goes offline.
Previously there was old cached data displayed, and now
the netowrk_type and block number gets updated to
"no connection" and "N/A".