Change the place where startup/shutdown actions are done due to log leak
Since Textual 0.80.0, all calls to textual logger done before active_app
ContextVar is set are being printed out to the terminal so logs are
visible after app is closed.

And this ContextVar is set only when app is running. We should do
startup and cleanup actions when TUI is already launched.

Still there is an issue which causes log leak from on_unmount:
https://github.com/Textualize/textual/issues/5091