Remove ParameterGroup
Rationale:
Main issue with this solution is that it is quite a hacky one because
Typer does not have solution for sharing options across multiple
commands and our solution had a side effect - we couldn't control
the param/option order displayed in help.

Looks like it's better to have some boilerplate than such a solution.

Because we have common options defined in one place, only parameters are
redundant. So it is not a blatant duplication, because the parameter definitions
are not duplicated, but only the parameters names.