Add gitlab pages redirect to default/
Previously the `GitLab Pages` button on repository
site was targeting https://clive-e34412.pages.syncad.com/

We do not have any documentation in this place because we store them in
different directories for each branch (feature branch and develop branch).
So for now we'll redirect to "default/".

In the future we'll probably need to introduce some directory for tag
so we'll also keep all of the production documentation available
(different versions for different tags).